requirements.txt
asgiref==3.2.10
boto3==1.14.43
botocore==1.17.43
Django==3.1
django-cors-headers==3.4.0
django-dynamic-model==0.2.0
django-environ==0.4.5
django-filter==2.3.0
django-storages==1.9.1
djangorestframework==3.11.1
docutils==0.15.2
jmespath==0.10.0
numpy==1.19.1
pandas==1.1.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
